improve z ordering to not rely on godot's global z-index
z-indexes are a bit broken for groups, in our case it's preferred to sort within the scene tree even if that's more expensive of an operation.  Render order in most models no not extensively change from frame to frame so it should result in minimal operational overhead versus the other workaround of wrapping models in viewports.